requirements.txt
requests
sqlmodel
structlog
colorama
PyYAML
typer
rich
celery
redis
psutil
beautifulsoup4
pytest
fastapi
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
markdown
weasyprint
markupsafe
uvicorn
types-requests
types-PyYAML
types-psutil
types-Markdown
types-python-jose
types-passlib
